Refactor DraftPage/SealedPage
Renamed CurrentPack/DefaultButton/HomePage to DraftPack/StartButton/DraftPage. Refactored DraftPage and SealedPage. Created SealedPool component. Added Loading indicator for sealed builds.
