+++ client/src/App.jsx
@@ -1,7 +1,7 @@
 import { CssBaseline, GlobalStyles, Typography, Box } from '@mui/material';
 import Navbar from './Navbar';
 import Footer from './Footer';
-import HomePage from './HomePage';
+import DraftPage from './DraftPage';
 import SealedPage from './SealedPage';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 
@@ -11,11 +11,10 @@ function App() {
       <CssBaseline />
       <GlobalStyles styles={{ body: { backgroundColor: 'rgba(26, 26, 26, 1)' } }} />
       <Navbar />
-
       <Box sx={{ display: 'flex', flexDirection: 'column', minHeight: '100vh' }}>
         <Box sx={{ flex: '1' }}>
           <Routes>
-            <Route path="/" element={<HomePage />} />
+            <Route path="/" element={<DraftPage />} />
             <Route path="/sealed" element={<SealedPage />} />
           </Routes>
         </Box>
