+++ client/src/StartButton.jsx
@@ -1,9 +1,9 @@
 import { Button } from '@mui/material';
 
-export default function DefaultButton({ draftStarted, children, onClick }) {
+export default function StartButton({ isLoading, children, onClick }) {
     const styles = {
-        DefaultButton: {
-            display: draftStarted ? 'none' : 'flex',
+        StartButton: {
+            display: isLoading ? 'none' : 'flex',
             backgroundColor: 'rgba(73, 73, 73, 1)',
             borderRadius: '20px',
             p: '0.6rem',
@@ -14,6 +14,6 @@ export default function DefaultButton({ draftStarted, children, onClick }) {
     }
 
     return (
-        <Button variant='contained' sx={styles.DefaultButton} onClick={onClick}>{children}</Button>
+        <Button variant='contained' sx={styles.StartButton} onClick={onClick}>{children}</Button>
     )
 }
